requirements.txt
httpx>=0.27.0
playwright>=1.40.0
pillow>=10.0.0
